+++ core/Action/RotateTo.cpp
@@ -2,7 +2,7 @@
 #include "..\e2dnode.h"
 
 
-e2d::RotateTo::RotateTo(double duration, double rotation)
+e2d::RotateTo::RotateTo(float duration, float rotation)
 	: RotateBy(duration, 0)
 {
 	_endVal = rotation;
