+++ core/Common/Size.cpp
@@ -6,7 +6,7 @@ e2d::Size::Size()
 	height = 0;
 }
 
-e2d::Size::Size(double width, double height)
+e2d::Size::Size(float width, float height)
 {
 	this->width = width;
 	this->height = height;
@@ -28,12 +28,12 @@ e2d::Size e2d::Size::operator-(Size const & size) const
 	return Size(width - size.width, height - size.height);
 }
 
-e2d::Size e2d::Size::operator*(double const & value) const
+e2d::Size e2d::Size::operator*(float const & value) const
 {
 	return Size(width * value, height * value);
 }
 
-e2d::Size e2d::Size::operator/(double const & value) const
+e2d::Size e2d::Size::operator/(float const & value) const
 {
 	return Size(width / value, height / value);
 }
